requirements/requirements.txt
Baker==1.3
CherryPy==3.7.0
Clay==2.12.5
colorama==0.3.3
diff-match-patch==20121119
dj-static==0.0.6
Django==1.7.11
django-admin-honeypot==0.4.0
django-appconf==1.0.1
django-ckeditor==5.0.2
django-colorfield==0.1.10
django-cors-headers==1.1.0
django-debug-toolbar==1.2.2
django-extensions==1.5.5
django-filebrowser==3.7.1
django-filebrowser-no-grappelli==3.6.2
django-geoposition==0.2.2
django-grappelli==2.8.1
django-import-export==0.2.8
django-jet==0.1.4
django-jsonview==0.5.0
django-localeurl==2.0.1
django-modeltranslation==0.11
django-mssql==1.7
django-nocaptcha-recaptcha==0.0.18
django-nose==1.3
django-paintstore==0.2
django-parler==1.2.1
django-preventconcurrentlogins==0.7
django-queryset-csv==0.3.1
django-ratelimit==0.6.0
django-recaptcha==1.0.4
django-reversion==1.8.7
django-richtextfield==1.0.1
django-rosetta==0.7.10
django-rosetta-grappelli==1.0.1
django-session-security==2.2.4
django-sslify==0.2.7
django-taggit==0.17.1
django-uuslug==1.1.6
django-watchman==0.5.0
ecdsa==0.11
Fabric==1.10.2
Flask==0.10.1
fpdf==1.7.2
gunicorn==19.4.5
html5lib==1.0b3
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
microsofttranslator==0.7
Moar==1.3.3
MySQL-python==1.2.5
nose==1.3.4
paramiko==1.15.2
pbkdf2==1.3
pep8==1.5.7
Pillow==2.9.0
pisa==3.0.33
polib==1.0.7
psycopg2==2.6
pycrypto==2.6.1
pyExcelerator==0.6.4.1
Pygments==2.0.2
pyjade==3.1.0
python-slugify==1.1.4
raven==5.5.0
reportlab==3.2.0
requests==2.5.1
singlemodeladmin==0.4
six==1.10.0
sqlparse==0.1.14
static3==0.6.1
tablib==0.10.0
Unidecode==0.4.18
unrar==0.3
Voodoo==1.4.1
Werkzeug==0.10.4
xlrd==0.9.4
xlwt==1.0.0
